Minor code cleanup
* Break large method into several smaller methods
* Remove unused configuration
* Respect `fatalWarnings = true`
