+++ scalafix-rules/src/main/scala/scalafix/internal/rule/CompilerException.scala
@@ -1,0 +1,4 @@
+package scalafix.internal.rule
+
+case class CompilerException(underlying: Throwable)
+    extends Exception("compiler crashed", underlying)
