+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/AnyRefWith.scala
