+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/EnumerationValue.scala
