+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/FinalVal.scala
