+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/ImmutableMap.scala
