+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/ImmutableSeq.scala
