+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/LowerBound.scala
