+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/NameConflict.scala
