+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/Rename.scala
