+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/ScalaPackage.scala
