+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/SeenFromTypeParams.scala
