+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/ThisTypeRef.scala
