+++ scalafix-tests/input/src/main/scala/test/explicitResultTypes/WidenSingleType.scala
