+++ scalafix-tests/input/src/main/scala/tests/ExplicitResultTypesArrayBuffer.scala
