+++ scalafix-tests/input/src/main/scala/tests/ExplicitResultTypesNil.scala
