+++ scalafix-tests/input/src/main/scala/tests/ExplicitResultTypesPartialFunction.scala
