+++ scalafix-tests/input/src/main/scala/tests/ExplicitResultTypesPrefix.scala
