+++ scalafix-tests/input/src/main/scala/tests/ExplicitResultTypesRefinement.scala
