+++ scalafix-tests/input/src/main/scala/tests/ExplicitResultTypesSupermethod.scala
