+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/AnyRefWith.scala
