+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/EnumerationValue.scala
