+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/FinalVal.scala
