+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/ImmutableMap.scala
