+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/ImmutableSeq.scala
