+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/LowerBound.scala
