+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/NameConflict.scala
