+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/Rename.scala
