+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/ScalaPackage.scala
