+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/SeenFromTypeParams.scala
