+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/ThisTypeRef.scala
