+++ scalafix-tests/output/src/main/scala/test/explicitResultTypes/WidenSingleType.scala
