+++ scalafix-tests/output/src/main/scala/tests/ExplicitResultTypesArrayBuffer.scala
