+++ scalafix-tests/output/src/main/scala/tests/ExplicitResultTypesBug.scala
