+++ scalafix-tests/output/src/main/scala/tests/ExplicitResultTypesNil.scala
