+++ scalafix-tests/output/src/main/scala/tests/ExplicitResultTypesPartialFunction.scala
