+++ scalafix-tests/output/src/main/scala/tests/ExplicitResultTypesPrefix.scala
