+++ scalafix-tests/output/src/main/scala/tests/ExplicitResultTypesRefinement.scala
