+++ scalafix-tests/output/src/main/scala/tests/ExplicitResultTypesSupermethod.scala
